// src/browser/app/src/workflow_contract.generated.ts
export interface RfDetrPresetContract {
  readonly presetName: string;
  readonly displayName: string;
}
// ...
export const RF_DETR_PRESET_OPTIONS = [
  {
    "displayName": "RF-DETR Nano",
    "presetName": "rf-detr-nano"
  },
  {
    "displayName": "RF-DETR Small",
    "presetName": "rf-detr-small"
  },
  {
    "displayName": "RF-DETR Medium",
    "presetName": "rf-detr-medium"
  },
  {
    "displayName": "RF-DETR Large",
    "presetName": "rf-detr-large"
  },
  {
    "displayName": "RF-DETR Seg Nano",
    "presetName": "rf-detr-seg-nano"
  },
  {
    "displayName": "RF-DETR Seg Small",
    "presetName": "rf-detr-seg-small"
  },
  {
    "displayName": "RF-DETR Seg Medium",
    "presetName": "rf-detr-seg-medium"
  },
  {
    "displayName": "RF-DETR Seg Large",
    "presetName": "rf-detr-seg-large"
  },
  {
    "displayName": "RF-DETR Seg XLarge",
    "presetName": "rf-detr-seg-xlarge"
  },
  {
    "displayName": "RF-DETR Seg XXLarge",
    "presetName": "rf-detr-seg-xxlarge"
  }
] as const satisfies readonly RfDetrPresetContract[];